+++ Stonks/src/portfolio/model/FlexiblePortfoliosModel.java
@@ -22,7 +22,7 @@ public class FlexiblePortfoliosModel extends AbstractPortfolioModel {
   }
 
   @Override
-  public void addStocksToPortfolio(String tickerSymbol, Long quantity,
+  public void addStocksToPortfolio(String tickerSymbol, Double quantity,
       int portfolioId, LocalDate date) {
     if (quantity < 0) {
       throw new IllegalArgumentException("Invalid quantity");
@@ -42,7 +42,7 @@ public class FlexiblePortfoliosModel extends AbstractPortfolioModel {
   }
 
   @Override
-  public void sellStockFromPortfolio(String tickerSymbol, Long quantity, int portfolioId,
+  public void sellStockFromPortfolio(String tickerSymbol, Double quantity, int portfolioId,
       LocalDate date) {
     if (quantity < 0) {
       throw new IllegalArgumentException("Invalid quantity");
@@ -106,7 +106,7 @@ public class FlexiblePortfoliosModel extends AbstractPortfolioModel {
     return this.transactionFee;
   }
 
-  protected AbstractPortfolio createPortfolio(Map<IStock, Long> stockQty, LocalDate date) {
+  protected AbstractPortfolio createPortfolio(Map<IStock, Double> stockQty, LocalDate date) {
     return new FlexiblePortfolio(this.stockService, stockQty, this.transactionFee, date);
   }
